Acordion Quarkus

+++ java/sgcs/src/main/java/org/cruzeirodosul/dao/DebitoDao.java
@@ -1,5 +1,6 @@
 package org.cruzeirodosul.dao;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import javax.inject.Singleton;
@@ -11,7 +12,7 @@ import io.quarkus.hibernate.orm.panache.PanacheRepository;
 @Singleton
 public class DebitoDao implements PanacheRepository<Debito> {
 
-	public List<Debito> findByIdCadastro(Long idCadastro) {
+	public List<Debito> findByIdCadastro(BigInteger idCadastro) {
 
 		return find("idcadastro", idCadastro).list();
