+++ java/sgcs/src/main/java/org/cruzeirodosul/resource/DebitoResource.java
@@ -1,5 +1,6 @@
 package org.cruzeirodosul.resource;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import javax.inject.Inject;
@@ -21,7 +22,7 @@ public class DebitoResource {
 	@GET
 	@Path("{idDesbravador}")
 	@Produces(MediaType.APPLICATION_JSON)
-	public List<Debito> buscarDebitoDesbravador(@PathParam Long idDesbravador) {
+	public List<Debito> buscarDebitoDesbravador(@PathParam BigInteger idDesbravador) {
 		return debitoDao.findByIdCadastro(idDesbravador);
 
 	}
